pdf_chat_django/requirements.txt
django==5.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.5